Command for saving deps into a file
This is an escape hatch to the JDK without having to have a really verbose CLI command inspired by jresolve.

It introduces a new subcommand `dep save`, which generates `.dep.*` files with JDK CLI flags and the links to dependencies inside `target/deps` or actually copies the jars if used with `--materialize` flad.

This also provides a mechanism for veles to not have to resolve the entire dependency graph on each run.