Integrated tests into compile command
The command has now two new "phases". It uses JUnit
to run unit and integration tests and uses Jacoco
to generate coverage reports.

In order to make test classpath work with JdtLS,
integration with Maven was needed. Veles now
generates pom.xml when running the lsp command,
which makes it compatible with other editors
and IDEs, not only with JdtLS. This also
means that the export command is not needed
anymore.

Summary:
- Unit tests that run after compilation
- Integration tests that run after everything is compiled
- Code coverage reports with Jacoco
- Options to filter the tests, skip slow tests, and filter in code with 'only' tag
- Generation of pom.xml for the custom project structure
- Removal of the export command