Moved --native-reach to start and run commands (#15)
The flag was misplaced as part of compile command.
It functions essentially the same as Leyden, and
I think it is better to have it away from compilation.
This flag essentially runs java commands with the same
flags, just uses Graal's JDK and attaches the agent.
It is now also possible to run other packaged outputs
with it and it isn't just the equivalent of the run
command anymore.